backend/requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
bcrypt==4.1.3
black==26.1.0
blinker==1.9.0
boto3==1.42.42
botocore==1.42.42
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
databricks-sdk==0.89.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
ecdsa==0.19.1
email-validator==2.3.0
emergentintegrations==0.1.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.20.3
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.62.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.4.0
idna==3.11
implicit==0.7.2
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jq==1.11.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
librt==0.7.8
litellm==1.80.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.4
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
motor==3.3.1
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
namex==0.1.0
numpy==2.4.2
oauthlib==3.3.1
openai==1.99.9
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
prettytable==3.17.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.5.0
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
scikit-learn==1.8.0
scikit-surprise==1.1.4
scipy==1.17.0
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlparse==0.5.5
starlette==0.37.2
stripe==14.3.0
surprise==0.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.25.0
wandb==0.25.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==15.0.1
Werkzeug==3.1.5
wrapt==2.1.1
yarl==1.22.0
zipp==3.23.0
